requirements.txt
streamlit>=1.0.0
pydeck>=0.7.0
geopandas>=0.10.0
shapely>=1.7.0
requests>=2.25.0